feat: preserve directory structure for API uploads
